+++ ParallaxGenLib/include/NIFUtil.hpp
@@ -4,6 +4,7 @@
 #include <NifFile.hpp>
 #include <Shaders.hpp>
 #include <array>
+#include <tuple>
 
 #define NUM_TEXTURE_SLOTS 9
 
@@ -20,6 +21,40 @@ enum class TextureSlots : unsigned int {
   Unused = 8
 };
 
+enum class PGTextureType {
+  DIFFUSE,
+  NORMAL,
+  MODELSPACENORMAL,
+  EMISSIVE,
+  SKINTINT,
+  SUBSURFACE,
+  HEIGHT,
+  CUBEMAP,
+  ENVIRONMENTMASK,
+  COMPLEXMATERIAL,
+  RMAOS,
+  TINT,
+  INNERLAYER,
+  COATNORMAL,
+  BACKLIGHT,
+  SPECULAR
+};
+
+auto getTextureTypeStr(const PGTextureType &Type) -> std::string;
+
+struct PGTexture {
+  std::filesystem::path Path;
+  PGTextureType Type;
+};
+
+auto getDefaultTextureType(const TextureSlots &Slot) -> PGTextureType;
+
+auto loadNIFFromBytes(const std::vector<std::byte> &NIFBytes) -> nifly::NifFile;
+
+auto getTexSuffixMap() -> std::map<std::wstring, std::tuple<TextureSlots, PGTextureType>>;
+
+auto getDefaultsFromSuffix(const std::filesystem::path &Path) -> std::tuple<TextureSlots, PGTextureType>;
+
 // shader helpers
 auto setShaderType(nifly::NiShader *NIFShader, const nifly::BSLightingShaderPropertyShaderType &Type,
                    bool &Changed) -> void;
@@ -27,32 +62,32 @@ auto setShaderFloat(float &Value, const float &NewValue, bool &Changed) -> void;
 auto setShaderVec2(nifly::Vector2 &Value, const nifly::Vector2 &NewValue, bool &Changed) -> void;
 
 // Shader flag helpers
-auto hasShaderFlag(nifly::BSLightingShaderProperty *NIFShaderBSLSP,
+auto hasShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP,
                    const nifly::SkyrimShaderPropertyFlags1 &Flag) -> bool;
-auto hasShaderFlag(nifly::BSLightingShaderProperty *NIFShaderBSLSP,
+auto hasShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP,
                    const nifly::SkyrimShaderPropertyFlags2 &Flag) -> bool;
-auto setShaderFlag(nifly::BSLightingShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags1 &Flag,
+auto setShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags1 &Flag,
                    bool &Changed) -> void;
-auto setShaderFlag(nifly::BSLightingShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags2 &Flag,
+auto setShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags2 &Flag,
                    bool &Changed) -> void;
-auto clearShaderFlag(nifly::BSLightingShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags1 &Flag,
+auto clearShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags1 &Flag,
                      bool &Changed) -> void;
-auto clearShaderFlag(nifly::BSLightingShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags2 &Flag,
+auto clearShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags2 &Flag,
                      bool &Changed) -> void;
-auto configureShaderFlag(nifly::BSLightingShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags1 &Flag,
+auto configureShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags1 &Flag,
                          const bool &Enable, bool &Changed) -> void;
-auto configureShaderFlag(nifly::BSLightingShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags2 &Flag,
+auto configureShaderFlag(nifly::BSShaderProperty *NIFShaderBSLSP, const nifly::SkyrimShaderPropertyFlags2 &Flag,
                          const bool &Enable, bool &Changed) -> void;
 
 // Texture slot helpers
+auto setTextureSlot(nifly::NifFile *NIF, nifly::NiShape *NIFShape, const TextureSlots &Slot, const std::wstring &TexturePath, bool &Changed) -> void;
 auto setTextureSlot(nifly::NifFile *NIF, nifly::NiShape *NIFShape, const TextureSlots &Slot,
                     const std::string &TexturePath, bool &Changed) -> void;
-auto getTexBase(const std::string &TexPath, const std::vector<std::string> &Suffixes) -> std::string;
-auto getTexMatch(const std::string &Base,
-                 const std::map<std::string, std::filesystem::path> &SearchMap) -> std::filesystem::path;
+auto getTexBase(const std::filesystem::path &TexPath) -> std::wstring;
+auto getTexMatch(const std::wstring &Base,
+                 const std::map<std::wstring, PGTexture> &SearchMap) -> PGTexture;
 // Gets all the texture prefixes for a textureset. ie. _n.dds is removed etc. for each slot
-auto getSearchPrefixes(nifly::NifFile &NIF, nifly::NiShape *NIFShape,
-                       const std::vector<std::vector<std::string>> &Suffixes)
-    -> std::array<std::string, NUM_TEXTURE_SLOTS>;
+auto getSearchPrefixes(nifly::NifFile &NIF, nifly::NiShape *NIFShape)
+    -> std::array<std::wstring, NUM_TEXTURE_SLOTS>;
 
 } // namespace NIFUtil