+++ ParallaxGenLib/include/patchers/PatcherComplexMaterial.hpp
@@ -26,11 +26,11 @@ public:
                          ParallaxGenDirectory *PGD, ParallaxGenD3D *PGD3D);
 
   // check if complex material should be enabled on shape
-  auto shouldApply(nifly::NiShape *NIFShape, const std::array<std::string, NUM_TEXTURE_SLOTS> &SearchPrefixes,
+  auto shouldApply(nifly::NiShape *NIFShape, const std::array<std::wstring, NUM_TEXTURE_SLOTS> &SearchPrefixes,
                    bool &EnableResult, bool &EnableDynCubemaps,
-                   std::string &MatchedPath) const -> ParallaxGenTask::PGResult;
+                   std::wstring &MatchedPath) const -> ParallaxGenTask::PGResult;
 
   // enables complex material on a shape in a NIF
-  auto applyPatch(nifly::NiShape *NIFShape, const std::string &MatchedPath, const bool &ApplyDynCubemaps,
+  auto applyPatch(nifly::NiShape *NIFShape, const std::wstring &MatchedPath, const bool &ApplyDynCubemaps,
                   bool &NIFModified) const -> ParallaxGenTask::PGResult;
 };