+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed PBR bug when multilayer: false is defined
 - Fixed upgrade-shader not generating mipmaps
 - Fixed PBR bug with duplicate texture sets
+- New method for mapping textures to types
 
 ## [0.5.7] - 2025-09-05